feat: 添加甘肃案例2（2026-01退休，男，60岁3个月，兰州市）
- 案例68：1965-10出生，1988-07参工，2026-01退休
- 延迟退休3个月（1965-10出生→60岁3个月→2026-01退休）
- 建账时间2003-01（特殊！非全省统一1996-01）
- 视同14.42年，实际23.08年，累计37.5年
- 个人账户314079.50元，平均缴费指数2.5955（极高）
- 基础5222.01 + 个人2287.54 + 过渡3478.92 = 10988.47
- 全部4项验证通过，一分不差
- 关键发现：甘肃建账时间存在地区差异
- 修复引擎：parseInput中months字段从parseInt改为parseFloat
- 更新甘肃配置：2025年基数7746，添加60.25岁计发月数137.3

+++ engine/pension-engine.js
@@ -718,7 +718,7 @@ function parseInput(inputData) {
   const skipDelay = inputData.skipDelay === true
 
   // monthsInput: 用户可显式指定计发月数（覆盖自动计算，用于提前退休验证场景）
-  const monthsInput = inputData.months != null ? parseInt(inputData.months) : null
+  const monthsInput = inputData.months != null ? parseFloat(inputData.months) : null
 
   return {
     name,
